requirements.txt
streamlit
google-generativeai
pydantic
python-dotenv
